backend/pyproject.toml
[build-system]
requires = ["setuptools>=42", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "chainsettle"
version = "0.1.0"
description = "Python Library for ChainSettle Nodes"
readme = "README.md"
requires-python = ">=3.7"
license = "MIT"

authors = [
    { name = "Brandyn Hamilton", email = "brandynham1120@gmail.com"}
]

dependencies = [
    "pandas",
    "numpy",
    "python-dotenv",
    "flask[async]",
    "web3",
    "plaid-python",
    "arweave-python-client",
    "paypal-server-sdk==1.0.0",
    "py-solc-x",
    "diskcache>=5.6.3",
    "click",
    "docusign-esign>=5.0.0",
]

[tool.setuptools.packages.find]
include = ["chainsettle", "chainsettle.*"]
